requirements.txt
pyyaml>=6.0.1
xmltodict>=0.13.0
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.1.0
black>=23.7.0
mypy>=1.5.1
bandit>=1.7.5
pydocstyle>=6.3.0
drawio-parser>=0.1.0 